Move `cfd_trading_change_notifier` to `controller`
