+++ lib/controller/cfd_trading_change_notifier.dart
